🎓 Study : R3F 학습 13 Shadow (Drei)
- [x] ✅ 📝 Docs : RM_13.md 문서 작성
