+++ README/RM_13.md
@@ -1,0 +1,47 @@
+---
+title: R3F의 shadow (Drei)
+description: shadow (Drei) 대한 설명 및 실습입니다.
+date: 2023-11-02T08:34:16.694Z
+preview: https://i.ytimg.com/vi/bLM-fleqfs4/maxresdefault.jpg
+draft: false
+tags:
+    - R3F
+categories:
+    - R3F
+---
+
+## 🎓 강의를 기반으로 학습하는 과정입니다
+
+> R3F 에대한 강의로 3D 웹 개발에 대하여 입문과정을 알려준다
+
+[![image](https://i.ytimg.com/vi/bLM-fleqfs4/maxresdefault.jpg)](https://www.youtube.com/watch?v=Sg6OcVxe64k&list=PLe6NQuuFBu7HUeJkowKRkLWwkdOlhwrje&index=13)
+
+## 📔 Shadow 종류
+
+| 명칭 | 모양 | 이미지 |
+| :-- | :-- | :-- |
+| [AccumulativeShadows](#📝-pointlightshadow) | 프레임 갯수만큼 별로 겹처 그리는 그림자 | [![image](./assets/AccumulativeShadows.gif)](https://pmndrs.github.io/drei-vanilla/?path=/story/shaders-accumulativeshadows--acc-shadow-story) |
+| [ContactShadows](#📝-directionallightshadow) | 평면으로 표현된 일반적인 그림자 입니다 | [![image](./assets/ContactShadows.gif)](https://drei.pmnd.rs/?path=/docs/staging-contactshadows--docs) |
+
+> [!WARRING]  
+> `Canvas`의 shadow 속성을 설정해주어야 한다
+
+### 📝 AccumulativeShadows
+
+PointLight에 대한 쉐도우
+
+### 👀 Preview(AccumulativeShadows)
+
+![image](./assets/AccumulativeShadows.gif)
+
+🔗 Link : <https://pmndrs.github.io/drei-vanilla/?path=/story/shaders-accumulativeshadows--acc-shadow-story>
+
+### 📝 ContactShadows
+
+DirectionalLight에 대한 쉐도우
+
+### 👀 Preview(ContactShadows)
+
+![image](./assets/ContactShadows.gif)
+
+🔗 Link : <https://drei.pmnd.rs/?path=/docs/staging-contactshadows--docs>
